Move to using MeshSpecs for all PrimitiveModelStreams, and refactor SphereModelStream to subdivide initial vertex positions, for speed & simplicity

+++ mpp/include/mpp/GridModelStream.h
@@ -10,6 +10,6 @@ namespace mpp
 	{
 	public:
 
-		GridModelStream(ResourceManager* resourceMgr, mesh::MeshSpecification const& meshSpec, std::string const& material, double width, double depth, int dimX, int dimZ);
+		GridModelStream(ResourceManager* resourceMgr, mesh::MeshSpecification const& meshSpec, std::string const& material, double width, double depth, size_t dimX, size_t dimZ);
 	};
 }
